% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidGseaTestsUrl}
\alias{AcidGseaTestsUrl}
\title{AcidGSEA test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidGseaTestsUrl
}
\description{
AcidGSEA test data URL
}
\examples{
AcidGseaTestsUrl
}
\keyword{internal}
