% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,list,FGSEAList-method}
\title{Force an Object to Belong to a Class}
\value{
Object of new class.
}
\description{
Coerce an object to a given class.
}
\section{Description}{

    \code{as(object)}
      returns the version of this object coerced to be the given
      \code{Class}.  When used in the replacement form on the left of
      an assignment, the portion of the object corresponding to
      \code{Class} is replaced by \code{value}.

      The operation of \code{as()} in either form depends on the
      definition of coerce methods.  Methods are defined automatically
      when the two classes are related by inheritance; that is, when
      one of the classes is a subclass of the other.

      Coerce methods are also predefined for basic classes (including all
      the types of vectors, functions and a few others).

      Beyond these two sources of methods, further methods are defined
      by calls to the \code{\link{setAs}} function.  See that
      documentation also for details of how coerce methods work. Use
      \code{showMethods(coerce)} for a list of all currently defined methods, as in the
      example below.

 
  
}

\section{Basic Coercion Methods}{

  Methods are pre-defined for coercing any object to one of the basic
  datatypes.  For example, \code{as(x, "numeric")} uses the existing
  \code{as.numeric} function.  These and all other existing methods
  can be listed as shown in the example.

}

\examples{
# DESeqDataSet ====
x <- as(bcb_small, "DESeqDataSet")
names(S4Vectors::mcols(x))
class(x)
show(x)

# RangedSummarizedExperiment ====
x <- as(bcb_small, "RangedSummarizedExperiment")
slotNames(x)
show(x)

# SummarizedExperiment ====
# Coerce to RangedSummarizedExperiment first.
x <- as(bcb_small, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)
slotNames(x)
show(x)
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
\itemize{
\item \code{\link[methods:coerce]{methods::coerce()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[methods:as]{methods::as()}}.
}
}
