% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/collectionNames-methods.R
\name{collectionNames}
\alias{collectionNames}
\alias{collectionNames<-}
\alias{collectionNames,FgseaList-method}
\alias{collectionNames<-,FgseaList,character-method}
\title{Gene set collection names}
\usage{
collectionNames(object, ...)
collectionNames(object, ...) <- value

\S4method{collectionNames}{FgseaList}(object)

\S4method{collectionNames}{FgseaList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Gene set collection names
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
collectionNames(object)
}
