% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,FGSEAList,FGSEAList-method}
\title{Combining or merging different Bioconductor data structures}
\usage{
combine(x, y, ...)

\S4method{combine}{FGSEAList,FGSEAList}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
A single value of the same class as the most specific common ancestor
  (in class terms) of the input values. This will contain the
  appropriate combination of the data in the input values.
}
\description{
Combining or merging different Bioconductor data structures
}
\note{
Updated 2019-11-19.
}
\examples{
data(gsea)
x <- gsea
contrastNames(x) <- paste0("x_", contrastNames(x))
y <- gsea
contrastNames(y) <- paste0("y_", contrastNames(y))
c <- combine(x, y)
c
}
