% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastNames-methods.R
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames<-}
\alias{contrastNames,FGSEAList-method}
\alias{contrastNames<-,FGSEAList,character-method}
\title{Contrast names}
\usage{
contrastNames(object, ...)
contrastNames(object, ...) <- value

\S4method{contrastNames}{FGSEAList}(object)

\S4method{contrastNames}{FGSEAList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Contrast names
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
contrastNames(object)
}
