% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotEnrichment}
\alias{plotGSEATable}
\title{Defunct functions}
\usage{
plotEnrichment(...)

plotGSEATable(...)
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\keyword{internal}
