% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseq}
\alias{deseq}
\title{DESeqAnalysis example data}
\format{
An object of class \code{DESeqAnalysis} of length 1.
}
\usage{
deseq
}
\description{
Used to generate example \code{FGSEAList} object.
}
\examples{
data(deseq)
print(deseq)
}
\keyword{datasets}
