% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/enriched-methods.R
\docType{methods}
\name{enriched}
\alias{enriched}
\alias{enriched,FGSEAList-method}
\title{Enriched pathways from gene set collections}
\usage{
enriched(object, ...)

\S4method{enriched}{FGSEAList}(object, alpha = 0.05)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}
}
\value{
list.
Named list formatted as:
\enumerate{
\item Gene set collection (e.g. "h" from MSigDb).
\item Contrast (e.g. "dmso_r1881_vs_etoh").
\item Direction (e.g. "down" or "up").
}
}
\description{
Enriched pathways from gene set collections
}
\examples{
data(gsea)
enriched(gsea)
}
