% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea}
\alias{fgsea}
\alias{pfgsea}
\title{Fast Gene Set Enrichment Analysis}
\usage{
fgsea(stats, gmtFile, ...)

pfgsea(gmtFiles, statsList, ...)
}
\arguments{
\item{stats}{\code{numeric}. Gene stats. Test statistic is recommended by default,
but shrunken log2 fold changes are also acceptable. Don't input unshrunken
LFC values without standard error correction.}

\item{gmtFile}{\code{string}. MSigDB GMT file path.}

\item{...}{Additional arguments.}

\item{gmtFiles}{\code{character}. MSigDB GMT file paths.}

\item{statsList}{\code{list}. Gene stats list.}
}
\description{
Wrapper for \code{\link[fgsea:fgsea]{fgsea::fgsea()}} that loads the pathways automatically from a
GMT file, rather than requiring a separate call to \code{\link[fgsea:gmtPathways]{fgsea::gmtPathways()}}.
}
\section{Functions}{
\itemize{
\item \code{fgsea}: Returns a \code{data.frame}/\code{data.table} of results.

\item \code{pfgsea}: Parameterized variant. Returns a \code{list} of results.
}}

