% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FgseaList-class}
\alias{FgseaList-class}
\title{Fast GSEA list}
\value{
\code{FgseaList}.
}
\description{
Class containing parameterized fast GSEA results.
}
\note{
Updated 2022-08-16.
}
