% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,FGSEAList-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\details{
For some organisms, gene identifiers and gene names do not map 1:1 (e.g.
\emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument
here in the documentation for approaches that deal with this issue.
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
g2s <- Gene2Symbol(object)
print(g2s)
}
\seealso{
\code{\link[AcidGenomes:makeGene2Symbol]{makeGene2Symbol()}}.
}
