% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSet-methods.R
\name{geneSet}
\alias{geneSet}
\alias{geneSet,FGSEAList-method}
\title{Gene set}
\usage{
geneSet(object, ...)

\S4method{geneSet}{FGSEAList}(object, collection, set)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}
}
\value{
\code{character}.
}
\description{
Gene set
}
\note{
Updated 2020-09-16.
}
\examples{
data(fgsea)
geneSet(
    object = fgsea,
    collection = 1L,
    set = 1L
)
}
