% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSetNames-methods.R
\name{geneSetNames}
\alias{geneSetNames}
\alias{geneSetNames,FGSEAList-method}
\title{Gene set names}
\usage{
geneSetNames(object, ...)

\S4method{geneSetNames}{FGSEAList}(object, collection)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}
}
\value{
\code{character}.
}
\description{
Gene set names
}
\note{
Updated 2020-09-23.
}
\examples{
data(fgsea)
geneSetNames(object = fgsea, collection = "h")
}
