% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSetResults-methods.R
\name{geneSetResults}
\alias{geneSetResults}
\alias{geneSetResults,FGSEAList-method}
\title{Gene set results}
\usage{
geneSetResults(object, ...)

\S4method{geneSetResults}{FGSEAList}(object, contrast, collection, set)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gene set results
}
\note{
Updated 2021-03-16.
}
\examples{
data(fgsea)
geneSetResults(
    object = fgsea,
    contrast = "condition_B_vs_A",
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
