% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneToSymbol-methods.R
\name{GeneToSymbol}
\alias{GeneToSymbol}
\alias{GeneToSymbol,FgseaList-method}
\title{Gene-to-symbol mappings}
\usage{
GeneToSymbol(object, ...)

\S4method{GeneToSymbol}{FgseaList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{GeneToSymbol}.
}
\description{
Gene-to-symbol mappings
}
\details{
For some organisms, gene identifiers and gene names do not map 1:1 (e.g.
\emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument
here in the documentation for approaches that deal with this issue.
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
g2s <- GeneToSymbol(object)
print(g2s)
}
