% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsea}
\alias{gsea}
\title{FGSEAList example data}
\format{
An object of class \code{FGSEAList} of length 1.
}
\usage{
gsea
}
\description{
FGSEAList example data
}
\examples{
data(gsea)
print(gsea)
}
\keyword{datasets}
