% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/leadingEdge-methods.R
\name{leadingEdge}
\alias{leadingEdge}
\alias{leadingEdge,FGSEAList-method}
\title{Leading edge genes}
\usage{
leadingEdge(object, ...)

\S4method{leadingEdge}{FGSEAList}(object, contrast, collection, set)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Elements of gene set that correspond to leading edge.
}
\description{
Leading edge genes
}
\note{
Updated 2020-09-21.
}
\examples{
data(fgsea)
leadingEdge(
    object = fgsea,
    contrast = "condition_B_vs_A",
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
