% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdownTables-methods.R
\name{markdownTables}
\alias{markdownTables}
\alias{markdownTables,FGSEAList-method}
\title{Markdown tables}
\usage{
markdownTables(object, ...)

\S4method{markdownTables}{FGSEAList}(object, collection, n = 10L, headerLevel = 3L)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Markdown tables
}
\details{
Top tables of significantly enriched pathways.

Supports looping across multiple DEG results, and adds a Markdown header for
each contrast.
}
\note{
Updated 2022-05-24.
}
\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
alphaThreshold(object) <- 0.9
collection <- collectionNames(object)[[1L]]
markdownTables(
    object = object,
    collection = collection
)
}
