% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nesThreshold-methods.R
\name{nesThreshold}
\alias{nesThreshold}
\alias{nesThreshold<-}
\alias{nesThreshold,FgseaList-method}
\alias{nesThreshold<-,FgseaList,numeric-method}
\title{Normalized enrichment score (NES) threshold}
\usage{
nesThreshold(object, ...)
nesThreshold(object, ...) <- value

\S4method{nesThreshold}{FgseaList}(object)

\S4method{nesThreshold}{FgseaList,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
}
\description{
Normalized enrichment score (NES) threshold
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
nesThreshold(object) <- 0.1
nesThreshold(object)
}
