% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pathwayNames-methods.R
\docType{methods}
\name{pathwayNames}
\alias{pathwayNames}
\alias{pathwayNames<-}
\alias{pathwayNames,FGSEAList-method}
\alias{pathwayNames<-,FGSEAList,character-method}
\title{Pathway names}
\usage{
pathwayNames(object, ...)

pathwayNames(object, ...) <- value

\S4method{pathwayNames}{FGSEAList}(object)

\S4method{pathwayNames}{FGSEAList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Pathway names
}
\examples{
data(gsea)
pathwayNames(gsea)
}
