% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pfgsea-package}
\title{pfgsea}
\description{
Perform parameterized gene set enrichment analysis on multiple differential
expression contrasts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steinbaugh/pfgsea}
  \item Report bugs at \url{https://github.com/steinbaugh/pfgsea/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
