% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfgsea.R
\name{pfgsea}
\alias{pfgsea}
\title{Parameterized fast gene set enrichment analysis (GSEA)}
\usage{
pfgsea(rankedList, gmtFiles, nPerm = 1000L, minSize = 15L,
  maxSize = 500L, BPPARAM = bpparam())
}
\arguments{
\item{rankedList}{\code{RankedList}.
Ranked gene list.}

\item{gmtFiles}{\code{character}.
GMT file paths.}

\item{nPerm}{\code{integer(1)}.
Number of permutations.
Minimial possible nominal \emph{P} value is about 1/\code{nPerm}.}

\item{minSize}{\code{integer(1)}.
Minimal size of a gene set to test.
All pathways below the threshold are excluded.}

\item{maxSize}{\code{integer(1)}/\code{Inf}.
Maximal size of a gene set to test.
All pathways above the threshold are excluded.}

\item{BPPARAM}{BiocParallel parallelization parameter.}
}
\description{
Extends the functionality of \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.
}
\examples{
data(gsea)
rankedList <- metadata(gsea)[["rankedList"]]
gmtFiles <- metadata(gsea)[["gmtFiles"]]

# Copy example MSigDb files to `$HOME`.
file.copy(
    from = system.file(
        "extdata", "msigdb",
        package = "pfgsea",
        mustWork = TRUE
    ),
    to = "~",
    overwrite = FALSE,
    recursive = TRUE
)

x <- pfgsea(rankedList = rankedList, gmtFiles = gmtFiles)
print(x)
}
