% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{pfgseaTestsURL}
\alias{pfgseaTestsURL}
\title{pfgsea test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
pfgseaTestsURL
}
\description{
pfgsea test data URL
}
\examples{
pfgseaTestsURL
}
\keyword{internal}
