% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\name{plotCounts}
\alias{plotCounts}
\title{Plot counts}
\usage{
plotCounts(object, ...)
}
\arguments{
\item{collection}{\code{character(1)} or \code{integer(1)}.
Collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
For example, \code{"c1"} or \code{"h"} (for hallmark).}

\item{contrast}{\code{character(1)} or \code{integer(1)}.
DESeqResults contrast.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{DESeqAnalysis}{\code{DESeqAnalysis}.
Corresponding DESeq2 data used to perform GSEA.}

\item{samples}{\code{character} or \code{NULL}.
Specific samples to plot.
If \code{NULL} (default), include all samples.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\note{
Updated 2019-11-07.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
rownames <- head(rownames(object))
print(rownames)
g2s <- basejump::Gene2Symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

## Rownames, gene IDs, and gene names (symbols) are supported.
plotCounts(object, genes = geneIDs, style = "facet")
plotCounts(object, genes = geneNames, style = "wide")
}
\seealso{
\code{DESeq2::plotCounts()}.
}
