% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEnrichedUpset-methods.R
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\alias{plotEnrichedUpset,FgseaList-method}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)

\S4method{plotEnrichedUpset}{FgseaList}(object, collection, direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
alphaThreshold(object) <- 0.9
collection <- collectionNames(object)[[1L]]
plotEnrichedUpset(
    object = object,
    collection = collection
)
}
