% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGSEAHeatmap-methods.R
\name{plotGSEAHeatmap}
\alias{plotGSEAHeatmap}
\title{Plot normalized GSEA scores as a heatmap}
\usage{
plotGSEAHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Plot normalized GSEA scores as a heatmap
}
\note{
Updated 2019-11-13.
}
\examples{
data(deseq, package = "DESeqAnalysis")
convertToHuman(deseq)
}
