% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEATable.R
\name{plotGSEATable}
\alias{plotGSEATable}
\alias{plotGSEATables}
\title{Plot GSEA Table}
\usage{
plotGSEATable(results, stats, gmtFile, alpha, n = 5L)

plotGSEATables(resultsList, statsList, gmtFile, alpha, n = 5L,
  headerLevel = 3L)
}
\arguments{
\item{results}{\code{data.frame}/\code{data.table}. Unmodified \code{\link[fgsea:fgsea]{fgsea::fgsea()}} return.}

\item{stats}{\code{numeric}. Gene stats. Test statistic is recommended by default,
but shrunken log2 fold changes are also acceptable. Don't input unshrunken
LFC values without standard error correction.}

\item{gmtFile}{\code{string}. MSigDB GMT file path.}

\item{alpha}{\code{scalar numeric}. Alpha level.}

\item{n}{\code{scalar integer}. Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{resultsList}{\code{list}. Results list.}

\item{statsList}{\code{list}. Gene stats list.}

\item{headerLevel}{\code{scalar integer}. Markdown header level.}
}
\value{
\code{ggplot}.
}
\description{
Wrapper for \code{\link[fgsea:plotGseaTable]{fgsea::plotGseaTable()}}.
}
\section{Functions}{
\itemize{
\item \code{plotGSEATable}: Plot a table of the top up- and down-regulated
processes in the results returned from \code{\link[=fgsea]{fgsea()}}.

\item \code{plotGSEATables}: Parameterized. Loops across the result contrasts
and adds a Markdown header.
}}

