% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap-methods.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,FGSEAList-method}
\title{Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{FGSEAList}(
  object,
  DESeqAnalysis,
  contrast,
  contrastSamples = TRUE,
  collection,
  set,
  leadingEdge = TRUE,
  scale = "row",
  clusteringMethod = "ward.D2",
  clusterRows = TRUE,
  clusterCols = TRUE,
  color = getOption(x = "acid.heatmap.color", default = acidplots::blueYellow),
  breaks = seq(from = -2L, to = 2L, by = 0.25),
  legendBreaks = seq(from = -2L, to = 2L, by = 1L),
  ...
)
}
\arguments{
\item{object}{Object.}

\item{leadingEdge}{\code{logical(1)}.
Visualize only the leading edge genes returned by GSEA. If \code{FALSE}, plot
all genes in the gene set, which requires the external gene set file that
was originally used to run the analysis. This file is referenced internally
inside the object at \code{metadata(object)[["gmtFiles"]]}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Construct a simple heatmap.
}
\note{
Updated 2019-11-18.
}
