% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotLfc-methods.R
\name{plotLfc}
\alias{plotLfc}
\alias{plotLfc,FgseaList-method}
\title{Plot log2 fold change distributions}
\usage{
plotLfc(object, ...)

\S4method{plotLfc}{FgseaList}(
  object,
  contrast = NULL,
  collection,
  set,
  geom = c("boxplot", "boxplot"),
  points = TRUE,
  labels = list(title = "log2 fold change", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character} or \code{NULL}.
Contrast name.
If \code{NULL}, plot all contrasts.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{points}{\code{logical(1)}.
Show individual data points.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot log2 fold change distributions
}
\note{
Updated 2023-08-15.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
collection <- collectionNames(object)[[1L]]
set <- geneSetNames(object = object, collection = collection)[[1L]]
plotLfc(
    object = object,
    collection = collection,
    set = set
)
}
