% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNES-methods.R
\name{plotNES}
\alias{plotNES}
\alias{plotNES,FGSEAList-method}
\title{Plot normalized enrichment score}
\usage{
plotNES(object, ...)

\S4method{plotNES}{FGSEAList}(
  object,
  contrast,
  collection,
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot normalized enrichment score
}
\note{
Updated 2021-09-10.
}
\examples{
## FGSEAList ====
data(fgsea)
plotNES(
    object = fgsea,
    contrast = "condition_B_vs_A",
    collection = "h"
)
}
\seealso{
Inspired by example in Stephen Turner's
\href{https://stephenturner.github.io/deseq-to-fgsea/}{DESeq to fgsea guide}.
}
