% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGeneSetFiles.R
\name{prepareGeneSetFiles}
\alias{prepareGeneSetFiles}
\title{Prepare gene set files}
\usage{
prepareGeneSetFiles(dir)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory name containing MSigDb release.}
}
\description{
Prepare gene set files
}
\details{
For example, intended for MSigDb releases.
}
\note{
Updated 2021-03-04.
}
\examples{
dir <- system.file(
    "extdata",
    "msigdb",
    "7.0",
    "msigdb_v7.0_GMTs",
    package = "AcidGSEA",
    mustWork = TRUE
)
files <- prepareGeneSetFiles(dir)
print(files)
}
