% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerators.R
\docType{class}
\name{RankedList-class}
\alias{RankedList-class}
\alias{RankedList,matrix-method}
\alias{RankedList,DESeqAnalysis-method}
\alias{RankedList,FGSEAList-method}
\title{Ranked list of genes}
\usage{
\S4method{RankedList}{matrix}(object, value = "log2FoldChange")

\S4method{RankedList}{DESeqAnalysis}(object, value = c("stat",
  "log2FoldChange", "padj"))

\S4method{RankedList}{FGSEAList}(object)
}
\value{
\code{RankedList}.
}
\description{
Class containing parameterized ranked gene lists.
}
\note{
Updated 2019-07-24.
}
