% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RankedList-methods.R
\docType{methods}
\name{RankedList}
\alias{RankedList}
\alias{RankedList,DESeqAnalysis-method}
\alias{RankedList,FGSEAList-method}
\title{Prepare a ranked gene (stats) list for GSEA}
\usage{
RankedList(object, ...)

\S4method{RankedList}{DESeqAnalysis}(object, value = c("stat",
  "log2FoldChange", "padj"))

\S4method{RankedList}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{\code{character(1)}.
Value type to use for GSEA. Recommending \code{"stat"} by default, which
corresponds to the Wald test statistic for DESeq2. Shrunken log2 fold
change (\code{"log2FoldChange"}) or BH-adjusted \emph{P} values ("padj") are also
acceptible.}
}
\value{
\code{RankedList}.
}
\description{
Return a parameterized ranked list for each differential expression contrast.
}
\section{Gene symbol multi-mapping}{


Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the stat values using \code{mean()} internally.
}

\examples{
## x <- RankedList(object)
## names(x)
}
