% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/results-methods.R
\name{results}
\alias{results}
\alias{results,FGSEAList-method}
\title{Results}
\usage{
results(object, ...)

\S4method{results}{FGSEAList}(object, contrast, collection)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Results
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
contrast <- contrastNames(object)[[1L]]
collection <- collectionNames(object)[[1L]]
results(
    object = fgsea,
    contrast = contrast,
    collection = collection
)
}
