% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,FGSEAList-method}
\title{Show an object}
\usage{
\S4method{show}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\examples{
data(gsea)

## FGSEAList ====
show(gsea)
}
