% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statsList-methods.R
\docType{methods}
\name{statsList}
\alias{statsList}
\alias{statsList,DESeqAnalysis-method}
\title{Prepare stats list for GSEA}
\usage{
statsList(object, ...)

\S4method{statsList}{DESeqAnalysis}(object, value = c("stat",
  "log2FoldChange", "padj"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{\code{character(1)}.
Value type to use for GSEA. Recommending \code{"stat"} by default, which
corresponds to the Wald test statistic for DESeq2. Shrunken log2 fold
change (\code{"log2FoldChange"}) or BH-adjusted \emph{P} values ("padj") are also
acceptible.}
}
\description{
Return a stats list for each gene symbol.
}
\section{Gene symbol multi-mapping}{


Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the stat values using \code{mean()} internally.
}

\examples{
## x <- statsList(object)
## names(x)
}
