% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTable.R
\name{topTable}
\alias{topTable}
\alias{topTables}
\title{Top Table}
\usage{
topTable(results, alpha, n = 10L)

topTables(resultsList, alpha, n = 10L, headerLevel = 3L)
}
\arguments{
\item{results}{\code{data.frame}/\code{data.table}.
Unmodified \code{fgsea::fgsea} return.}

\item{alpha}{\code{scalar numeric}.
Alpha level.}

\item{n}{\code{scalar integer}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{resultsList}{\code{list}.
Results list.}

\item{headerLevel}{\code{scalar integer}.
Markdown header level.}
}
\value{
\code{data.frame}.
}
\description{
Top Table
}
\section{Functions}{
\itemize{
\item \code{topTable}: Return a top table of the most postiively associated processes, arranged by
\code{NES}. Must pass the \code{alpha} level cutoff.

\item \code{topTables}: Parameterized. Supports looping across multiple DEG results, and adds a
Markdown header for each contrast.
}}

