% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,FGSEAList-method}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)

\S4method{updateObject}{FGSEAList}(object, deseq = NULL, alphaThreshold = NULL, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{deseq}{\code{DESeqAnalysis} or \code{NULL}.
Update object by slotting \code{DESeqAnalysis}, if necessary.
This was added to object in 2020-09.}

\item{alphaThreshold}{\code{number(1)}.
Alpha level used for GSEA.
Note that this is not necessarily the alpha level used to generate
\code{DESeqResults} object.}

\item{...}{Additional arguments.}

\item{verbose}{\code{logical(1)}.
Whether information about the update should be reported.}
}
\value{
Modified object.
}
\description{
Update object
}
\note{
Updated 2022-03-11.
}
\examples{
data(fgsea)

## FGSEAList ====
updateObject(fgsea)
}
\author{
Michael Steinbaugh
}
