% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidGSEA-package}
\alias{AcidGSEA}
\alias{AcidGSEA-package}
\title{AcidGSEA}
\description{
Perform parameterized fast gene set enrichment analysis (GSEA) on multiple
differential expression contrasts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/acidgsea/}
  \item \url{https://github.com/acidgenomics/r-acidgsea/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidgsea/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})
}

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
