% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{acidgseaTestsURL}
\alias{acidgseaTestsURL}
\title{acidgsea test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
acidgseaTestsURL
}
\description{
acidgsea test data URL
}
\examples{
acidgseaTestsURL
}
\keyword{internal}
