% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectionNames-methods.R
\docType{methods}
\name{collectionNames}
\alias{collectionNames}
\alias{collectionNames<-}
\alias{collectionNames,FGSEAList-method}
\alias{collectionNames<-,FGSEAList,character-method}
\title{Gene set collection names}
\usage{
collectionNames(object, ...)

collectionNames(object, ...) <- value

\S4method{collectionNames}{FGSEAList}(object)

\S4method{collectionNames}{FGSEAList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Gene set collection names
}
\examples{
data(gsea)
collectionNames(gsea)
}
