% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,FGSEAList,FGSEAList-method}
\title{Combining or merging different Bioconductor data structures}
\usage{
combine(x, y, ...)

\S4method{combine}{FGSEAList,FGSEAList}(x, y)
}
\arguments{
\item{x, y}{Object.}

\item{...}{Additional arguments.}
}
\value{
A single value of the same class as the most specific common ancestor
  (in class terms) of the input values. This will contain the
  appropriate combination of the data in the input values.
}
\description{
The \code{combine} generic function handles methods for combining
  or merging different Bioconductor data structures.
  It should, given an arbitrary number of arguments of the same class
  (possibly by inheritance), combine them into a single instance in
  a sensible way (some methods may only combine 2 objects,
  ignoring \code{...} in the argument list; because Bioconductor
  data structures are complicated, check carefully that \code{combine}
  does as you intend).
}
\details{
There are two basic combine strategies. One is an intersection
  strategy. The returned value should only have rows (or columns) that
  are found in all input data objects. The union strategy says that the
  return value will have all rows (or columns) found in any one of the
  input data objects (in which case some indication of what to use for
  missing values will need to be provided).

  These functions and methods are currently under construction. Please
  let us know if there are features that you require.
}
\examples{
data(gsea)
x <- gsea
contrastNames(x) <- paste0("x_", contrastNames(x))
y <- gsea
contrastNames(y) <- paste0("y_", contrastNames(y))
c <- combine(x, y)
c
}
\seealso{
\itemize{
    \item \link[Biobase]{combine,AnnotatedDataFrame,AnnotatedDataFrame-method},
          \link[Biobase]{combine,AssayData,AssayData-method},
          \link[Biobase]{combine,MIAME,MIAME-method},
          and \link[Biobase]{combine,eSet,eSet-method} in the \pkg{Biobase}
          package for additional \code{combine} methods.

    \item \code{\link{merge}} for merging two data frames (or data-frame-like)
          objects.

    \item \code{\link[methods]{showMethods}} for displaying a summary of the
          methods defined for a given generic function.

    \item \code{\link[methods]{selectMethod}} for getting the definition of
          a specific method.

    \item \link{BiocGenerics} for a summary of all the generics defined
          in the \pkg{BiocGenerics} package.
  }
}
