% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,FGSEAList,FGSEAList-method}
\title{Combine multiple objects}
\usage{
combine(x, y, ...)

\S4method{combine}{FGSEAList,FGSEAList}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Combine multiple objects
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FGSEAList ====
x <- fgsea
y <- fgsea
contrastNames(x) <- paste0("x_", contrastNames(x))
contrastNames(y) <- paste0("y_", contrastNames(y))
c <- combine(x, y)
c
}
