% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastNames-methods.R
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames<-}
\alias{contrastNames,FgseaList-method}
\alias{contrastNames<-,FgseaList,character-method}
\title{Contrast names}
\usage{
contrastNames(object, ...)
contrastNames(object, ...) <- value

\S4method{contrastNames}{FgseaList}(object)

\S4method{contrastNames}{FgseaList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Contrast names
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
contrastNames(object)
}
