% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastNames-methods.R
\docType{methods}
\name{contrastNames}
\alias{contrastNames}
\alias{contrastNames<-}
\alias{contrastNames,FGSEAList-method}
\alias{contrastNames<-,FGSEAList,character-method}
\title{Contrast names}
\usage{
contrastNames(object, ...)

contrastNames(object, ...) <- value

\S4method{contrastNames}{FGSEAList}(object)

\S4method{contrastNames}{FGSEAList,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\description{
Contrast names
}
\examples{
data(gsea)
contrastNames(gsea)
}
