% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/convertToHuman-methods.R
\docType{methods}
\name{convertToHuman}
\alias{convertToHuman}
\alias{convertToHuman,DESeqAnalysis-method}
\title{Convert data set to use human orthologs}
\usage{
convertToHuman(object, ...)

\S4method{convertToHuman}{DESeqAnalysis}(object, map = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{map}{\code{DataFrame}, or \code{NULL}.
Ortholog mappings data frame returned by \code{\link[=matchHumanOrthologs]{matchHumanOrthologs()}}. Since
this function depends on the BioMart API and has a tendancy to time out,
we're allowing passthrough of a cached object here instead. If left \code{NULL},
then `\code{\link[=matchHumanOrthologs]{matchHumanOrthologs()}} will be called internally.}
}
\value{
Modified object.
Features (i.e. rownames) will be remapped to human genes.
}
\description{
Convert data set to use human orthologs
}
\examples{
data(deseq, package = "DESeqAnalysis")
convertToHuman(deseq)
}
