% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/enriched-methods.R
\docType{methods}
\name{enriched}
\alias{enriched}
\alias{enriched,FGSEAList-method}
\title{Enriched pathways}
\usage{
enriched(object, ...)

\S4method{enriched}{FGSEAList}(object, alpha = 0.05)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}
}
\value{
list.
Named list formatted as:
\enumerate{
\item pathway (e.g. "h")
\item contrast (e.g. "dmso_r1881_vs_etoh")
\item direction (e.g. "down")
}
}
\description{
Enriched pathways
}
\examples{
data(gsea)
enriched(gsea)
}
