% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/enrichedGeneSets-methods.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\alias{enrichedGeneSets,FgseaList-method}
\title{Enriched gene sets}
\usage{
enrichedGeneSets(object, ...)

\S4method{enrichedGeneSets}{FgseaList}(object, collection, direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
list.
Named list containing significant gene sets per contrast.
}
\description{
Enriched pathways from gene set collections.
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
alphaThreshold(object) <- 0.7
collection <- collectionNames(object)[[1L]]
enrichedGeneSets(
    object = object,
    collection = collection,
    direction = "up"
)
}
\seealso{
\itemize{
\item \code{DESeqAnalysis::deg()}.
}
}
