% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedGeneSets-methods.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\alias{enrichedGeneSets,FGSEAList-method}
\title{Enriched gene sets}
\usage{
enrichedGeneSets(object, ...)

\S4method{enrichedGeneSets}{FGSEAList}(object, collection, direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name, coresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
list.
Named list containing significant gene sets per contrast.
}
\description{
Enriched pathways from gene set collections.
}
\note{
Updated 2020-09-18.
}
\examples{
data(fgsea)

## FGSEAList ====
alphaThreshold(fgsea) <- 0.7
enrichedGeneSets(
    object = fgsea,
    collection = "h",
    direction = "up"
)
}
\seealso{
\itemize{
\item \code{DESeqAnalysis::deg()}.
}
}
