% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedGeneSets-methods.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\alias{enrichedGeneSets,FGSEAList-method}
\title{Enriched gene sets}
\usage{
enrichedGeneSets(object, ...)

\S4method{enrichedGeneSets}{FGSEAList}(
  object,
  collection,
  alpha = NULL,
  nesThreshold = NULL,
  direction = c("both", "up", "down")
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{nesThreshold}{\code{numeric(1)} or \code{NULL}.
NES cutoff threshold.
If left \code{NULL}, no cutoff will be applied.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
list.
Named list containing significant gene sets per contrast.
}
\description{
Enriched pathways from gene set collections.
}
\note{
Updated 2020-03-18.
}
\examples{
data(fgsea)
enrichedGeneSets(
    object = fgsea,
    collection = "h",
    alpha = 0.7,
    direction = "up"
)
}
\seealso{
\itemize{
\item \code{DESeqAnalysis::deg}.
}
}
