% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedGeneSets-methods.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\alias{enrichedGeneSets,FGSEAList-method}
\title{Enriched gene sets}
\usage{
enrichedGeneSets(object, ...)

\S4method{enrichedGeneSets}{FGSEAList}(
  object,
  collection,
  alpha = NULL,
  nesThreshold = NULL,
  direction = c("both", "up", "down"),
  flatten = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{nesThreshold}{\code{numeric(1)} or \code{NULL}.
NES cutoff threshold.
If left \code{NULL}, no cutoff will be applied.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{flatten}{\code{logical(1)}.
Flatten nested "up"/"down" directional enrichment vector subsets.
Recomended by default for UpSet plots.}

\item{...}{Additional arguments.}
}
\value{
list.
Named list formatted as:
\enumerate{
\item Gene set collection (e.g. "h" from MSigDb).
\item Contrast (e.g. "condition_B_vs_A").
\item Direction (e.g. "down" or "up").
}
}
\description{
Enriched pathways from gene set collections.
}
\note{
Updated 2020-03-18.
}
\examples{
data(gsea)
enrichedGeneSets(gsea, collection = "h", alpha = 0.6)
}
\seealso{
\itemize{
\item \code{DESeqAnalysis::deg}.
}
}
