% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedGeneSets-methods.R
\name{enrichedGeneSets}
\alias{enrichedGeneSets}
\alias{enrichedGeneSets,FGSEAList-method}
\title{Enriched pathways}
\usage{
enrichedGeneSets(object, ...)

\S4method{enrichedGeneSets}{FGSEAList}(object, collection, flatten = TRUE)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h}, \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{flatten}{\code{logical(1)}.
Flatten nested "up"/"down" directional enrichment vector subsets.
Recomended by default for UpSet plots.}

\item{...}{Additional arguments.}
}
\value{
list.
Named list formatted as:
\enumerate{
\item Gene set collection (e.g. "h" from MSigDb).
\item Contrast (e.g. "dmso_r1881_vs_etoh").
\item Direction (e.g. "down" or "up").
}
}
\description{
Enriched pathways from gene set collections
}
\note{
Updated 2019-08-28.
}
\examples{
data(gsea)
enrichedGeneSets(gsea, collection = "h")
}
