% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,FGSEAList-method}
\title{Export}
\usage{
export(x, ...)

\S4method{export}{FGSEAList}(x, name = NULL, dir = ".")
}
\arguments{
\item{x}{Object.}

\item{name}{Name.}

\item{dir}{Directory.}

\item{...}{Additional arguments.}
}
\description{
Export data out of R and write to disk.
}
\examples{
data(gsea)
export(gsea, dir = "example")
list.files(file.path("example", "gsea"))
unlink("example", recursive = TRUE)
}
