% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,FGSEAList-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{FGSEAList}(object, name = NULL, dir = ".")
}
\arguments{
\item{object}{Object.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{...}{Additional arguments.}
}
\value{
Invisibily return file paths, either as \code{list} or \code{character}.
}
\description{
Export data out of R and write to disk.
}
\note{
Updated 2019-08-28.
}
\examples{
data(gsea)
export(gsea, dir = "example")
sort(list.files(file.path("example", "gsea")))
unlink("example", recursive = TRUE)
}
