% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,FGSEAList,character,missingOrNULL-method}
\title{Export}
\usage{
export(object, con, format, ...)

\S4method{export}{FGSEAList,character,missingOrNULL}(
  object,
  con,
  format,
  geneSetResults = FALSE,
  compress = getOption(x = "acid.export.compress", default = FALSE),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)}.
Directory path.}

\item{format}{\emph{Not currently supported.}}

\item{geneSetResults}{\code{logical(1)} or \code{character}.
Export per-gene set expression including log fold change values generated
from DESeq2. Can be slow when processing all MSigDB collections, so
disabled by default. Alternatively, can declare specific collections to
process (see \code{collectionNames} for supported values).}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\note{
Updated 2022-08-31.
}
\section{On-disk structure}{


Example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{file.path("object", "mutant_vs_control", "c1.csv")
}\if{html}{\out{</div>}}

S4 object is currently structured by:
\enumerate{
\item Gene set (c1-c8, h).
\item Contrast.
}

The object was structured in this manner to flow with the R Markdown
template. However, when writing to disk, I think it makes more sense to
organize by:
\enumerate{
\item Contrast
\item Gene set.
}
}

\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
con <- AcidBase::tempdir2()
out <- export(
    object = object,
    con = con
)
print(out)
AcidBase::unlink2(con)
}
