% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fgsea}
\alias{fgsea}
\title{FgseaList example data}
\format{
An object of class \code{FgseaList} of length 1.
}
\usage{
data(fgsea)
}
\description{
FgseaList example data
}
\note{
Updated 2021-06-29.
}
\examples{
data(fgsea)
print(fgsea)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
