% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fgsea}
\alias{fgsea}
\title{FGSEAList example data}
\format{
An object of class \code{function} of length 1.
}
\usage{
fgsea
}
\description{
FGSEAList example data
}
\examples{
data(fgsea)
print(fgsea)
}
\keyword{datasets}
