% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea}
\alias{fgsea}
\alias{pfgsea}
\title{Fast Gene Set Enrichment Analysis}
\usage{
fgsea(stats, gmtFile, nperm, minSize = 1, maxSize = Inf, nproc = 0,
  gseaParam = 1, BPPARAM = NULL)

pfgsea(statsList, gmtFiles, ...)
}
\arguments{
\item{stats}{\code{numeric}.
Gene stats. Test statistic is recommended by default, but shrunken log2
fold changes are also acceptable. Don't input unshrunken LFC values without
standard error correction.}

\item{gmtFile}{\code{character(1)}.
MSigDB GMT file path.}

\item{statsList}{\code{list}.
Gene stats list.}

\item{gmtFiles}{\code{character}.
MSigDB GMT file paths.}

\item{...}{Additional arguments.}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
The columns are the following:
\itemize{
 \item pathway -- name of the pathway as in `names(pathway)`;
 \item pval -- an enrichment p-value;
 \item padj -- a BH-adjusted p-value;
 \item ES -- enrichment score, same as in Broad GSEA implementation;
 \item NES -- enrichment score normalized to mean enrichment of random samples of the same size;
 \item nMoreExtreme` -- a number of times a random gene set had a more
     extreme enrichment score value;
 \item size -- size of the pathway after removing genes not present in `names(stats)`.
 \item leadingEdge -- vector with indexes of leading edge genes that drive the enrichment, see \url{http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading}.
}
}
\description{
Wrapper for \code{fgsea::fgsea} that loads the pathways automatically from a
GMT file, rather than requiring a separate call to \code{fgsea::gmtPathways}.
}
\section{Functions}{
\itemize{
\item \code{fgsea}: Returns a \code{data.frame}/\code{data.table} of results.

\item \code{pfgsea}: Parameterized variant. Returns a \code{list} of results.
}}

