% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/FGSEAList-methods.R
\docType{class}
\name{FGSEAList-class}
\alias{FGSEAList-class}
\alias{FGSEAList,DESeqAnalysis-method}
\title{Fast GSEA list}
\usage{
\S4method{FGSEAList}{DESeqAnalysis}(object, ...)
}
\value{
\code{FGSEAList}.
}
\description{
Class containing parameterized fast GSEA results.
}
\section{Methods (by generic)}{
\itemize{
\item \code{FGSEAList}: Arguments pass through to \code{RankedList} method.
}}

\note{
Updated 2020-09-23.
}
