% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FGSEAList-class}
\alias{FGSEAList-class}
\alias{FGSEAList}
\title{Fast GSEA list}
\description{
Class containing parameterized fast GSEA results.
}
\note{
Updated 2019-07-24.
}
\author{
Michael Steinbaugh
}
