% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FGSEAList-class}
\alias{FGSEAList-class}
\title{Fast GSEA list}
\value{
\code{FGSEAList}.
}
\description{
Class containing parameterized fast GSEA results.
}
\note{
Updated 2019-11-19.
}
