% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/FGSEAList-methods.R
\name{FGSEAList}
\alias{FGSEAList}
\alias{FGSEAList,DESeqAnalysis-method}
\alias{FGSEAList,DESeqResults-method}
\alias{FGSEAList,RankedList-method}
\title{Parameterized fast gene set enrichment analysis (GSEA)}
\usage{
FGSEAList(object, ...)

\S4method{FGSEAList}{DESeqAnalysis}(
  object,
  keyType = c("geneName", "ensemblId", "entrezId", "geneId", "rowname"),
  value = c("stat", "log2FoldChange"),
  proteinCodingOnly = FALSE,
  ...
)

\S4method{FGSEAList}{DESeqResults}(
  object,
  keyType = c("geneName", "ensemblId", "entrezId", "geneId", "rowname"),
  value = c("stat", "log2FoldChange"),
  rowRanges,
  proteinCodingOnly = FALSE,
  ...
)

\S4method{FGSEAList}{RankedList}(
  object,
  geneSetFiles,
  nPerm = 1000L,
  minSize = 15L,
  maxSize = 500L,
  alphaThreshold = 0.05
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{keyType}{`character(1).
Gene identifier format:
\itemize{
\item \code{"entrezId"}: Entrez identifiers (e.g. \code{7157}).
\item \code{"geneName"}: Gene names (a.k.a. symbols; e.g. \code{"TP53"}).
}}

\item{value}{\code{character(1)}.
Value type to use for GSEA ranked list.

Currently supported:
\enumerate{
\item \code{stat}: Wald test statistic. This column is returned by \code{results()}
but is removed in \code{DESeq2::lfcShrink()} return, currently.
\item \code{log2FoldChange}: Shrunken log2 fold change. Note that this option
requires \code{DESeq2::lfcShrink()} return to be slotted.
\item \code{padj}: Adjusted \emph{P} value. This don't provide directional ranks, but
is offered as a legacy option. Not generally recommended.
}}

\item{proteinCodingOnly}{\code{logical(1)}.
Restrict to protein coding genes only.}

\item{rowRanges}{\code{GenomicRanges} or \code{GenomicRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{geneSetFiles}{\code{character}.
Gene set file paths (i.e. GMT files).
MSigDB files are recommended by default.}

\item{nPerm}{\code{integer(1)}.
Number of permutations.
Minimial possible nominal \emph{P} value is about 1/\code{nPerm}.}

\item{minSize}{\code{integer(1)}.
Minimal size of a gene set to test.
All pathways below the threshold are excluded.}

\item{maxSize}{\code{integer(1)}/\code{Inf}.
Maximal size of a gene set to test.
All pathways above the threshold are excluded.}

\item{alphaThreshold}{\code{numeric(1)}.
Alpha level cutoff.
Stored internally in \code{\link[=alphaThreshold]{alphaThreshold()}}.
Applied only to plots and enriched gene set exports, but does not affect
the actual GSEA enrichment calculation.}
}
\value{
\code{FGSEAList}.
}
\description{
Extends the functionality of \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.
}
\note{
Updated 2022-04-27.
}
\examples{
data(deseq, package = "DESeqAnalysis")

## DESeqAnalysis ====
object <- deseq
geneSetFiles <- prepareGeneSetFiles(
    dir = system.file(
        "extdata",
        "msigdb",
        "7.0",
        "msigdb_v7.0_GMTs",
        package = "AcidGSEA"
    )
)
fgsea <- FGSEAList(
    object = object,
    geneSetFiles = geneSetFiles
)
print(fgsea)
}
