% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,RankedList-method}
\alias{Gene2Symbol,FGSEAList-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{RankedList}(object)

\S4method{Gene2Symbol}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\note{
Updated 2020-09-23.
}
\examples{
data(fgsea)
Gene2Symbol(fgsea)
}
\seealso{
\code{\link[AcidGenomes:makeGene2Symbol]{makeGene2Symbol()}}.
}
