% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSet-methods.R
\name{geneSet}
\alias{geneSet}
\alias{geneSet,FgseaList-method}
\title{Gene set}
\usage{
geneSet(object, ...)

\S4method{geneSet}{FgseaList}(object, collection, set)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Gene set
}
\note{
Updated 2022-04-27.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
collection <- collectionNames(object)[[1L]]
set <- geneSetNames(object = object, collection = collection)[[1L]]
x <- geneSet(
    object = object,
    collection = collection,
    set = set
)
head(x)
}
