% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetResults-methods.R
\name{geneSetResults}
\alias{geneSetResults}
\alias{geneSetResults,FGSEAList-method}
\title{Gene set results}
\usage{
geneSetResults(object, ...)

\S4method{geneSetResults}{FGSEAList}(object, DESeqAnalysis, contrast, collection, set)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gene set results
}
\note{
Updated 2020-09-21.
}
\examples{
data(fgsea, deseq)
geneSetResults(
    object = fgsea,
    DESeqAnalysis = deseq,
    contrast = "condition_B_vs_A",
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
