% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSetResults-methods.R
\name{geneSetResults}
\alias{geneSetResults}
\title{Gene set results}
\usage{
geneSetResults(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Gene set results
}
\note{
Updated 2020-09-21.
}
\examples{
data(fgsea)
geneSet(
    object = fgsea,
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
