% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nesThreshold-methods.R
\name{nesThreshold}
\alias{nesThreshold}
\alias{nesThreshold<-}
\alias{nesThreshold,FGSEAList-method}
\alias{nesThreshold<-,FGSEAList,numeric-method}
\title{Normalized enrichment score (NES) threshold}
\usage{
nesThreshold(object, ...)

nesThreshold(object, ...) <- value

\S4method{nesThreshold}{FGSEAList}(object)

\S4method{nesThreshold}{FGSEAList,numeric}(object) <- value
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
Normalized enrichment score (NES) threshold
}
\note{
Updated 2020-08-05.
}
\examples{
data(fgsea)

## FGSEAList ====
nesThreshold(fgsea) <- 0.1
nesThreshold(fgsea)
}
