% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{geneSet}{\code{character(1)}.
Gene set name. Can obtain using \code{names()} on \code{FGSEAList} object.}

\item{gmtFile}{\code{character(1)}.
MSigDB GMT file path.}

\item{gmtFiles}{\code{character}.
MSigDB GMT file paths.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{object}{Object.}

\item{pathways}{\code{character}.
Pathways.}

\item{results}{\code{data.frame}/\code{data.table}.
Unmodified \code{fgsea::fgsea} return.}

\item{resultsList}{\code{list}.
Results list.}

\item{stats}{\code{numeric}.
Gene stats. Test statistic is recommended by default, but shrunken log2
fold changes are also acceptable. Don't input unshrunken LFC values without
standard error correction.}

\item{statsList}{\code{list}.
Gene stats list.}

\item{statsList}{\code{list}.
Gene stats list.}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
