% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{DESeqAnalysis}{\code{DESeqAnalysis}.
Corresponding DESeq2 data used to perform GSEA.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{contrastSamples}{\code{logical(1)}.
Only visualize the samples defined in the contrast.}

\item{geneSetFiles}{\code{character}.
Gene set file paths (i.e. GMT files).
MSigDB files are recommended by default.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}

\item{keyType}{`character(1).
Gene identifier format:
\itemize{
\item \code{"geneName"}: Gene names (a.k.a. symbols; e.g. \code{"TP53"}).
\item \verb{"ensemblGeneId}: Ensembl gene identifiers (e.g. \code{"ENSG00000000003"}).
\item \code{"ncbiGeneId"}: NCBI (Entrez) gene identifiers (e.g. \code{7157}).
}}

\item{leadingEdge}{\code{logical(1)}.
Visualize only the leading edge genes returned by GSEA.
If \code{FALSE}, plot all genes in the gene set.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{nesThreshold}{\code{numeric(1)} or \code{NULL}.
NES cutoff threshold.
If left \code{NULL}, no cutoff will be applied.}

\item{object}{Object.}

\item{pathways}{\code{character}.
Pathways.}

\item{results}{\code{data.frame}/\code{data.table}.
Unmodified \code{fgsea::fgsea} return.}

\item{resultsList}{\code{list}.
Results list.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{stats}{\code{numeric}.
Gene stats. Test statistic is recommended by default, but shrunken log2
fold changes are also acceptable. Don't input unshrunken LFC values without
standard error correction.}

\item{statsList}{\code{list}.
Gene stats list.}

\item{theme}{\code{theme}/\code{gg}.
ggplot2 theme.}

\item{value}{\code{character(1)}.
Value type to use for GSEA ranked list.

Currently supported:
\enumerate{
\item \code{stat}: Wald test statistic. This column is returned by \code{results()}
but is removed in \code{DESeq2::lfcShrink()} return, currently.
\item \code{log2FoldChange}: Shrunken log2 fold change. Note that this option
requires \code{DESeq2::lfcShrink()} return to be slotted.
\item \code{padj}: Adjusted \emph{P} value. This don't provide directional ranks, but
is offered as a legacy option. Not generally recommended.
}}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
