% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pfgsea-package}
\title{pfgsea}
\description{
Perform \strong{p}arameterized \strong{fast} \strong{g}ene \strong{s}et \strong{e}nrichment
\strong{a}nalysis on multiple differential expression contrasts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pfgsea.acidgenomics.com/pfgsea/}
  \item \url{https://github.com/acidgenomics/pfgsea/}
  \item Report bugs at \url{https://github.com/acidgenomics/pfgsea/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
