% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,FGSEAList-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{FGSEAList}(
  object,
  DESeqAnalysis,
  contrast,
  collection,
  set,
  n = 12L,
  contrastSamples = TRUE,
  line = "mean",
  ...
)
}
\arguments{
\item{object}{Object.}

\item{DESeqAnalysis}{\code{DESeqAnalysis}.
Corresponding DESeq2 data used to perform GSEA.}

\item{contrast}{\code{character(1)} or \code{integer(1)}.
DESeqResults contrast.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
For example, \code{"c1"} or \code{"h"} (for hallmark).}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{contrastSamples}{\code{logical(1)}.
Only visualize the samples defined in the contrast.}

\item{line}{\code{character(1)}.
Include average (median, mean, or geometric mean) line for each interesting
group. Disabled by default and if samples are colored by sample name.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\note{
Updated 2020-01-09.
}
