% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEnrichedGeneSets-methods.R
\name{plotEnrichedGeneSets}
\alias{plotEnrichedGeneSets}
\alias{plotEnrichedGeneSets,FGSEAList-method}
\title{Plot enriched gene sets}
\usage{
plotEnrichedGeneSets(object, ...)

\S4method{plotEnrichedGeneSets}{FGSEAList}(
  object,
  collection,
  direction = c("both", "up", "down"),
  n = 10L,
  headerLevel = 3L
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot enriched gene sets
}
\note{
Updated 2020-09-21.
}
\examples{
data(fgsea)
alphaThreshold(fgsea) <- 0.9
plotEnrichedGeneSets(fgsea, collection = "h", n = 1L)
}
\seealso{
\code{\link[=plotGeneSet]{plotGeneSet()}}.
}
