% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedGeneSets-methods.R
\name{plotEnrichedGeneSets}
\alias{plotEnrichedGeneSets}
\alias{plotEnrichedGeneSets,FGSEAList-method}
\title{Plot enriched gene sets}
\usage{
plotEnrichedGeneSets(object, ...)

\S4method{plotEnrichedGeneSets}{FGSEAList}(
  object,
  collection,
  alpha = NULL,
  nesThreshold = NULL,
  direction = c("both", "up", "down"),
  n = 10L,
  headerLevel = 3L,
  theme = acidplots::acid_theme_light()
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{nesThreshold}{\code{numeric(1)} or \code{NULL}.
NES cutoff threshold.
If left \code{NULL}, no cutoff will be applied.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{theme}{\code{theme}/\code{gg}.
ggplot2 theme.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot enriched gene sets
}
\note{
Updated 2020-03-18.
}
\examples{
## This requires MSigDB to be installed at `${HOME}`.
## > data(fgsea)
## > plotEnrichedGeneSets(fgsea, collection = "h", alpha = 0.9, n = 1L)
}
\seealso{
\code{\link[fgsea:plotEnrichment]{fgsea::plotEnrichment()}}.
}
