% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedUpset-methods.R
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\alias{plotEnrichedUpset,FGSEAList-method}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)

\S4method{plotEnrichedUpset}{FGSEAList}(
  object,
  collection,
  alpha = NULL,
  nesThreshold = NULL,
  direction = c("both", "up", "down")
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{nesThreshold}{\code{numeric(1)} or \code{NULL}.
NES cutoff threshold.
If left \code{NULL}, no cutoff will be applied.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\note{
Updated 2020-03-18.
}
\examples{
data(gsea)
plotEnrichedUpset(gsea, collection = "h", alpha = 0.9)
}
