% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedUpset-methods.R
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\alias{plotEnrichedUpset,FGSEAList-method}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)

\S4method{plotEnrichedUpset}{FGSEAList}(object, collection, direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name, coresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\note{
Updated 2020-08-05.
}
\examples{
data(fgsea)

## FGSEAList ====
alphaThreshold(fgsea) <- 0.9
plotEnrichedUpset(fgsea, collection = "h")
}
