% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEnrichedUpset-methods.R
\docType{methods}
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\alias{plotEnrichedUpset,FGSEAList-method}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)

\S4method{plotEnrichedUpset}{FGSEAList}(object, alpha = 0.05)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}
}
\value{
Graphical output.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\examples{
plotEnrichedUpset(gsea)
}
