% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichedUpset-methods.R
\docType{methods}
\name{plotEnrichedUpset}
\alias{plotEnrichedUpset}
\alias{plotEnrichedUpset,FGSEAList-method}
\title{UpSet plot of directional enriched pathway intersections across contrasts}
\usage{
plotEnrichedUpset(object, ...)

\S4method{plotEnrichedUpset}{FGSEAList}(object, collection, alpha = 0.05)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h}, \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{...}{Additional arguments.}
}
\value{
Graphical output.
}
\description{
UpSet plot of directional enriched pathway intersections across contrasts
}
\examples{
data(gsea)
plotEnrichedUpset(gsea, collection = "h")
}
