% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEnrichment-methods.R
\docType{methods}
\name{plotEnrichment}
\alias{plotEnrichment}
\alias{plotEnrichment,FGSEAList-method}
\title{Plot enrichment}
\usage{
plotEnrichment(object, ...)

\S4method{plotEnrichment}{FGSEAList}(object, geneSet, alpha = 0.05,
  n = 10L, headerLevel = 3L, theme = acid_theme_light())
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geneSet}{\code{character(1)}.
Gene set name. Can obtain using \code{names()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}

\item{theme}{\code{theme}/\code{gg}.
ggplot2 theme.}
}
\value{
\code{ggplot}.
}
\description{
Plot enrichment
}
\examples{
data(gsea)
plotEnrichment(gsea, geneSet = "h", n = 1L)
}
\seealso{
\code{\link[fgsea:plotEnrichment]{fgsea::plotEnrichment()}}.
}
