% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEnrichment-methods.R
\docType{methods}
\name{plotEnrichment}
\alias{plotEnrichment}
\alias{plotEnrichment,FGSEAList-method}
\title{Plot enrichment}
\usage{
plotEnrichment(object, ...)

\S4method{plotEnrichment}{FGSEAList}(object, geneSet, alpha = 0.05,
  n = 5L, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geneSet}{\code{character(1)}.
Gene set name. Can obtain using \code{names()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}
}
\value{
\code{ggplot}.
}
\description{
Wrapper for \code{\link[fgsea:plotEnrichment]{fgsea::plotEnrichment()}} that enables easy plotting of multiple
pathways of interest in a single call.
}
\seealso{
\code{\link[fgsea:plotEnrichment]{fgsea::plotEnrichment()}}.
}
