% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\alias{plotEnrichments}
\title{Plot Enrichment}
\usage{
plotEnrichment(pathways, stats, gmtFile, headerLevel = 3L)

plotEnrichments(resultsList, statsList, gmtFile, alpha, n = 5L,
  headerLevel = 3L)
}
\arguments{
\item{pathways}{\code{character}. Pathways.}

\item{stats}{\code{numeric}. Gene stats. Test statistic is recommended by default,
but shrunken log2 fold changes are also acceptable. Don't input unshrunken
LFC values without standard error correction.}

\item{gmtFile}{\code{string}. MSigDB GMT file path.}

\item{headerLevel}{\code{scalar integer}. Markdown header level.}

\item{resultsList}{\code{list}. Results list.}

\item{statsList}{\code{list}. Gene stats list.}

\item{alpha}{\code{scalar numeric}. Alpha level.}

\item{n}{\code{scalar integer}. Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}
}
\value{
\code{ggplot}.
}
\description{
Wrapper for \code{\link[fgsea:plotEnrichment]{fgsea::plotEnrichment()}} that enables easy plotting of multiple
pathways of interest in a single call.
}
\section{Functions}{
\itemize{
\item \code{plotEnrichment}: Parameterized. Returns multiple \code{ggplot}s,
with a Markdown header for each pathway.

\item \code{plotEnrichments}: Parameterized. Loops across the contrasts and
returns an additional Markdown header, on top of the \code{\link[=plotEnrichment]{plotEnrichment()}}
return per pathway.
}}

