% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGeneSet-methods.R
\name{plotGeneSet}
\alias{plotGeneSet}
\alias{plotGeneSet,FgseaList-method}
\title{Plot gene set enrichment}
\usage{
plotGeneSet(object, ...)

\S4method{plotGeneSet}{FgseaList}(
  object,
  collection,
  contrast,
  set,
  colors = c(line = "black", min = AcidPlots::purpleOrange(n = 2L)[[1L]], max =
    AcidPlots::purpleOrange(n = 2L)[[2L]], ticks = "black", yintercept = "black")
)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{colors}{\code{character(5)}.
Named character color vector indicating:
\itemize{
\item Line color.
\item Minimum (negative) threshold line color.
\item Maximum (positive) threhsold color.
\item Y intercept at origin color.
\item Ticks color.
}}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot gene set enrichment
}
\note{
Updated 2023-08-15.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
contrast <- contrastNames(object)[[1L]]
collection <- collectionNames(object)[[1L]]
set <- geneSetNames(object = object, collection = collection)[[1L]]
plotGeneSet(
    object = object,
    collection = collection,
    contrast = contrast,
    set = set
)
}
\seealso{
Modified version of \code{fgsea::plotEnrichment()}.
}
