% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGeneSet.R
\name{plotGeneSet}
\alias{plotGeneSet}
\alias{plotGeneSet,FGSEAList-method}
\title{Plot gene set enrichment}
\usage{
plotGeneSet(object, ...)

\S4method{plotGeneSet}{FGSEAList}(object, collection, contrast, set)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{collection}{\code{character(1)} or \code{integer(1)}.
Gene set collection name or position, corresponding to values defined in
\code{\link[base:names]{names()}}.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{contrast}{\code{character(1)} or \code{integer(1)}.
DESeqResults contrast.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}
}
\description{
Plot gene set enrichment
}
\note{
Updated 2020-07-22.
}
