% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEATable-methods.R
\docType{methods}
\name{plotGSEATable}
\alias{plotGSEATable}
\alias{plotGSEATable,FGSEAList-method}
\title{Plot GSEA enrichment table}
\usage{
plotGSEATable(object, ...)

\S4method{plotGSEATable}{FGSEAList}(object, collection, n = 10L,
  headerLevel = 3L)
}
\arguments{
\item{object}{Object.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h}, \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot GSEA enrichment table
}
\examples{
data(gsea)
plotGSEATable(gsea, collection = "h")
}
\seealso{
\code{\link[fgsea:plotGseaTable]{fgsea::plotGseaTable()}}.
}
