% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap-methods.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,FGSEAList-method}
\title{Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{FGSEAList}(
  object,
  contrast,
  contrastSamples = TRUE,
  collection,
  set,
  leadingEdge = FALSE,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{contrastSamples}{\code{logical(1)}.
Only visualize the samples defined in the contrast.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{leadingEdge}{\code{logical(1)}.
Visualize only the leading edge genes returned by GSEA.
If \code{FALSE}, plot all genes in the gene set.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Heatmap
}
\note{
Updated 2020-09-23.
}
\examples{
data(fgsea)
plotHeatmap(
    object = fgsea,
    contrast = "condition_B_vs_A",
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
