% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap-methods.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,FgseaList-method}
\title{Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{FgseaList}(
  object,
  contrast,
  contrastSamples = TRUE,
  collection,
  set,
  leadingEdge = FALSE,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{contrastSamples}{\code{logical(1)}.
Only visualize the samples defined in the contrast.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FgseaList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{leadingEdge}{\code{logical(1)}.
Visualize only the leading edge genes returned by GSEA.
If \code{FALSE}, plot all genes in the gene set.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Heatmap
}
\note{
Updated 2021-03-16.
}
\examples{
data(fgsea)

## FgseaList ====
object <- fgsea
contrast <- contrastNames(object)[[1L]]
collection <- collectionNames(object)[[1L]]
set <- geneSetNames(object = object, collection = collection)[[1L]]
plotHeatmap(
    object = object,
    contrast = contrast,
    collection = collection,
    set = set
)
}
