% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap-methods.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,FGSEAList-method}
\title{Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{FGSEAList}(
  object,
  DESeqAnalysis,
  contrast,
  contrastSamples = TRUE,
  collection,
  set,
  leadingEdge = TRUE,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{leadingEdge}{\code{logical(1)}.
Visualize only the leading edge genes returned by GSEA. If \code{FALSE}, plot
all genes in the gene set, which requires the external gene set file that
was originally used to run the analysis. This file is referenced internally
inside the object at \code{metadata(object)[["gmtFiles"]]}.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Construct a simple heatmap.
}
\note{
Updated 2019-11-19.
}
