% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotLFC-methods.R
\name{plotLFC}
\alias{plotLFC}
\alias{plotLFC,FGSEAList-method}
\title{Plot log2 fold change distributions}
\usage{
plotLFC(object, ...)

\S4method{plotLFC}{FGSEAList}(
  object,
  contrast = NULL,
  collection,
  set,
  geom = c("boxplot", "boxplot"),
  points = TRUE,
  color = getOption(x = "acid.color.discrete", default = getOption(x =
    "ggplot2.discrete.colour", default = AcidPlots::scale_color_synesthesia_d()))
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character} or \code{NULL}.
Contrast name.
If \code{NULL}, plot all contrasts.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{set}{\code{character(1)}.
Gene set name, in a defined \code{collection}.
For example, \code{"HALLMARK_ADIPOGENESIS"}.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{points}{\code{logical(1)}.
Show individual data points.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot log2 fold change distributions
}
\note{
Updated 2020-10-01.
}
\examples{
data(fgsea)
plotLFC(
    object = fgsea,
    collection = "h",
    set = "HALLMARK_P53_PATHWAY"
)
}
