% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNES-methods.R
\name{plotNES}
\alias{plotNES}
\alias{plotNES,FGSEAList-method}
\title{Plot normalized enrichment score}
\usage{
plotNES(object, ...)

\S4method{plotNES}{FGSEAList}(
  object,
  contrast,
  collection,
  n = Inf,
  flip = getOption(x = "acid.flip", default = TRUE),
  labels = list(title = TRUE, subtitle = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{n}{\code{integer(1)} or \code{Inf}.
Number of gene sets (regardless of direction) to include in plot.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot normalized enrichment score
}
\details{
Only plots gene sets that pass adjusted \emph{P} value cutoff, defined by
\code{alphaThreshold}.
}
\note{
Updated 2023-08-15.
}
\examples{
data(fgsea)

## FGSEAList ====
object <- fgsea
alphaThreshold(object) <- 0.9
contrast <- contrastNames(object)[[1L]]
collection <- collectionNames(object)[[1L]]
plotNES(
    object = object,
    contrast = contrast,
    collection = collection,
    n = 10L
)
}
\seealso{
Inspired by example in Stephen Turner's
\href{https://stephenturner.github.io/deseq-to-fgsea/}{DESeq to fgsea guide}.
}
