% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNES-methods.R
\name{plotNES}
\alias{plotNES}
\alias{plotNES,FGSEAList-method}
\title{Plot normalized enrichment scores of a collection}
\usage{
plotNES(object, ...)

\S4method{plotNES}{FGSEAList}(
  object,
  contrast,
  collection,
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = acidplots::scale_fill_synesthesia_d())),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{contrast}{\code{character(1)}.
Contrast name.}

\item{collection}{\code{character(1)}.
Gene set collection name.
Typically refers to \code{h} (hallmark), \code{c1}-\code{c7} collections from MSigDb.
Can obtain using \code{collectionNames()} on \code{FGSEAList} object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}
}
\value{
\code{ggplot}.
}
\description{
Plot normalized enrichment scores of a collection
}
\note{
Updated 2020-09-17.
}
\examples{
data(fgsea)
plotNES(
    object = fgsea,
    contrast = "condition_B_vs_A",
    collection = "h"
)
}
\seealso{
Inspired by example in Stephen Turner's
\href{https://stephenturner.github.io/deseq-to-fgsea/}{DESeq to fgsea guide}.
}
