% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGeneSetFiles.R
\name{prepareGeneSetFiles}
\alias{prepareGeneSetFiles}
\title{Prepare gene set files}
\usage{
prepareGeneSetFiles(
  dir,
  keyType = c("geneName", "entrezId"),
  ext = "gmt",
  recursive = FALSE
)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory name containing MSigDb release.}

\item{keyType}{`character(1).
Gene identifier format:
\itemize{
\item \code{"entrezId"}: Entrez identifiers (e.g. \code{7157}).
\item \code{"geneName"}: Gene names (a.k.a. symbols; e.g. \code{"TP53"}).
}}

\item{ext}{\code{character(1)}.
Gene set file extension.
Case insensitive.}

\item{recursive}{\code{logical(1)}.
Whether to search for gene set files recursively in \code{dir} argument.}
}
\description{
Prepare gene set files
}
\details{
Intended primarily to match GMT files from MSigDb.
}
\note{
Updated 2021-10-20.
}
\examples{
dir <- system.file(
    "extdata",
    "msigdb",
    "7.0",
    "msigdb_v7.0_GMTs",
    package = "AcidGSEA",
    mustWork = TRUE
)
files <- prepareGeneSetFiles(dir, keyType = "geneName")
print(files)
}
\seealso{
\itemize{
\item https://www.gsea-msigdb.org/gsea/msigdb/
\item https://www.ncbi.nlm.nih.gov/gene/
}
}
