% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/RankedList-methods.R
\docType{class}
\name{RankedList-class}
\alias{RankedList-class}
\alias{RankedList,matrix-method}
\title{Ranked list of genes}
\usage{
\S4method{RankedList}{matrix}(object, value = "log2FoldChange")
}
\description{
Class containing parameterized ranked gene lists.
}
\section{Methods (by generic)}{
\itemize{
\item \code{RankedList}: Contrast matrix, containing either log2 fold changes, or (Wald) test
statistic values. Useful when aggregating reuslts from multiple differential
expression analyses that can't be containerized into a DESeqAnalysis object.
}}

\author{
Michael Steinbaugh
}
