% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RankedList-class}
\alias{RankedList-class}
\title{Ranked list of genes}
\description{
Class containing parameterized ranked gene lists.
}
\seealso{
\code{\link[=RankedList]{RankedList()}}.
}
\author{
Michael Steinbaugh
}
