% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllGenerators.R
\name{RankedList}
\alias{RankedList}
\title{Ranked list of genes}
\usage{
RankedList(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{\code{character(1)}.
Value type to use for GSEA. Currently supported:
\enumerate{
\item \code{stat}: Wald test statistic. This column is returned by \code{results()}
but is removed in \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return, currently.
\item \code{log2FoldChange}: Shrunken log2 fold change. Note that this option
requires \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return to be slotted.
\item \code{padj}: Adjusted \emph{P} value. This don't provide directional ranks, but
is offered as a legacy option. Not generally recommended.
}}
}
\value{
\code{RankedList}.
}
\description{
Class containing parameterized ranked gene lists.
}
\note{
Updated 2019-08-28.
}
\section{Gene symbol multi-mapping}{


Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the stat values using \code{mean()} internally.
}

\examples{
## DESeqAnalysis ====
data(deseq, package = "DESeqAnalysis")
x <- RankedList(deseq)
print(x)

## FGSEAList ====
data(gsea, package = "pfgsea")
x <- RankedList(gsea)
print(x)

## matrix ====
data(matrix_lfc, package = "acidtest")
x <- RankedList(matrix_lfc)
print(x)
}
