% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RankedList-methods.R
\name{RankedList}
\alias{RankedList}
\alias{RankedList,matrix-method}
\alias{RankedList,DESeqResults-method}
\alias{RankedList,DESeqAnalysis-method}
\alias{RankedList,FGSEAList-method}
\title{Ranked list of genes}
\usage{
RankedList(object, ...)

\S4method{RankedList}{matrix}(object, value = "log2FoldChange")

\S4method{RankedList}{DESeqResults}(object, gene2symbol, value = c("stat", "log2FoldChange", "padj"))

\S4method{RankedList}{DESeqAnalysis}(
  object,
  value = c("stat", "log2FoldChange", "padj"),
  BPPARAM = BiocParallel::bpparam()
)

\S4method{RankedList}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{\code{character(1)}.
Value type to use for GSEA. Currently supported:
\enumerate{
\item \code{stat}: Wald test statistic. This column is returned by \code{results()}
but is removed in \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return, currently.
\item \code{log2FoldChange}: Shrunken log2 fold change. Note that this option
requires \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return to be slotted.
\item \code{padj}: Adjusted \emph{P} value. This don't provide directional ranks, but
is offered as a legacy option. Not generally recommended.
}}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{RankedList}.
}
\description{
Class containing parameterized ranked gene lists.
}
\note{
Updated 2020-03-17.
}
\section{Gene symbol multi-mapping}{


Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the stat values using \code{mean()} internally.
}

\examples{
## DESeqAnalysis ====
data(deseq, package = "DESeqAnalysis")
x <- RankedList(deseq)
print(x)

## FGSEAList ====
data(fgsea, package = "acidgsea")
x <- RankedList(fgsea)
print(x)

## matrix ====
data(matrix_lfc, package = "acidtest")
x <- RankedList(matrix_lfc)
print(x)
}
