% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RankedList-methods.R
\name{RankedList}
\alias{RankedList}
\alias{RankedList,DataFrame-method}
\alias{RankedList,DESeqAnalysis-method}
\alias{RankedList,DESeqResults-method}
\alias{RankedList,FGSEAList-method}
\title{Ranked list of genes}
\usage{
RankedList(object, ...)

\S4method{RankedList}{DataFrame}(object, rowRanges, keyType, value, proteinCodingOnly)

\S4method{RankedList}{DESeqAnalysis}(
  object,
  keyType = c("geneName", "geneId", "entrezId"),
  value = c("stat", "log2FoldChange"),
  proteinCodingOnly = FALSE
)

\S4method{RankedList}{DESeqResults}(
  object,
  rowRanges,
  keyType,
  value = c("stat", "log2FoldChange"),
  proteinCodingOnly = FALSE
)

\S4method{RankedList}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{rowRanges}{\code{GenomicRanges} or \code{GenomicRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{keyType}{`character(1).
Gene identifier format:
\itemize{
\item \code{"entrezId"}: Entrez identifiers (e.g. \code{7157}).
\item \code{"geneName"}: Gene names (a.k.a. symbols; e.g. \code{"TP53"}).
}}

\item{value}{\code{character(1)}.
Value type to use for GSEA ranked list.

Currently supported:
\enumerate{
\item \code{stat}: Wald test statistic. This column is returned by \code{results()}
but is removed in \code{DESeq2::lfcShrink()} return, currently.
\item \code{log2FoldChange}: Shrunken log2 fold change. Note that this option
requires \code{DESeq2::lfcShrink()} return to be slotted.
\item \code{padj}: Adjusted \emph{P} value. This don't provide directional ranks, but
is offered as a legacy option. Not generally recommended.
}}

\item{proteinCodingOnly}{\code{logical(1)}.
Restrict to protein coding genes only.}
}
\value{
\code{RankedList}.
}
\description{
Class containing parameterized ranked gene lists.
}
\note{
Updated 2022-04-27.
}
\section{Gene symbol multi-mapping}{


Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the values using \code{mean()} internally.
}

\examples{
## DESeqAnalysis ====
data(deseq, package = "DESeqAnalysis")
x <- RankedList(deseq)
print(x)

## FGSEAList ====
data(fgsea, package = "AcidGSEA")
x <- RankedList(fgsea)
print(x)
}
