% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bpparam}
\alias{acid_theme_light}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{acidplots}{\code{\link[acidplots]{acid_theme_light}}}

  \item{BiocParallel}{\code{\link[BiocParallel]{bpparam}}}
}}

