% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R, R/reexports.R,
%   R/topTables-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme_paperwhite}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{export}}, \code{\link[bioverbs]{topTables}}}

  \item{minimalism}{\code{\link[minimalism]{theme_paperwhite}}}
}}

