% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/results-methods.R
\name{results}
\alias{results}
\title{Results}
\usage{
results(object, ...)

results(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Varies, depending on the method.
}
\description{
Results
}
\note{
Updated 2020-09-21.
}
\examples{
data(fgsea)

## FGSEAList ====
results(fgsea)
}
