% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,FGSEAList-method}
\title{Show an object}
\usage{
\S4method{show}{FGSEAList}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Show an object
}
\note{
Updated 2020-09-22.
}
\examples{
data(fgsea)

## FGSEAList ====
show(fgsea)
}
