% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables-methods.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables,FGSEAList-method}
\title{Top tables}
\usage{
\S4method{topTables}{FGSEAList}(object, geneSet, alpha = 0.05, n = 5L,
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{geneSet}{\code{character(1)}.
Gene set name. Can obtain using \code{names()} on \code{FGSEAList} object.}

\item{alpha}{\code{numeric(1)}.
Alpha level.}

\item{n}{\code{integer(1)}.
Number of significant processes to include.
Bidirectional; will include \code{n} up- and down-regulated processes.}

\item{headerLevel}{\code{integer(1)}.
Markdown header level.}
}
\value{
Markdown output.
}
\description{
Top tables of significantly enriched pathways.
}
\details{
Supports looping across multiple DEG results, and adds a Markdown header for
each contrast.
}
